Experimenting with using HashSet instead of List
Seeing what it looks like in the code. I didn't like the deleting an
item from a list and adding it back to it's head as a way of using a
list as a set so trying with an actual set in Elixir. Code looks fine,
not sure about my test or how it would feel to external callers to get
and Elixir datastructure and not a list.